requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
jupyter>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
